Moving patientIdentifierMerge method into HibernatePationDAO. Fixed formatting
