--- api/src/main/java/org/openmrs/api/db/PatientIdentifierMergeDAO.java
@@ -1,23 +1,0 @@
-/**
- * This Source Code Form is subject to the terms of the Mozilla Public License,
- * v. 2.0. If a copy of the MPL was not distributed with this file, You can
- * obtain one at http://mozilla.org/MPL/2.0/. OpenMRS is also distributed under
- * the terms of the Healthcare Disclaimer located at http://openmrs.org/license.
- *
- * Copyright (C) OpenMRS Inc. OpenMRS is a registered trademark and the OpenMRS
- * graphic logo is a trademark of OpenMRS Inc.
- */
-package org.openmrs.api.db;
-
-import org.openmrs.PatientIdentifierType;
-
-/**
- * Database methods for the PatientIdentifierMerge
- * 
- * @see org.openmrs.api.context.Context
- */
-public interface PatientIdentifierMergeDAO {
-	
-	public void mergePatientIdentifier(PatientIdentifierType main, PatientIdentifierType toBeMerged) throws DAOException;
-	
-}
