+++ api/src/main/java/org/openmrs/api/db/hibernate/HibernatePatientDAO.java
@@ -995,4 +995,37 @@ public class HibernatePatientDAO implements PatientDAO {
     	sessionFactory.getCurrentSession().save(allergy);
     	return allergy;
     }
+    
+    /**
+     * @param main  main patient identifier type to be merged into
+     * @param toBeMerged a list of patient identifier types that are merged to main
+	 */
+	@Override
+	public void mergePatientIdentifier(PatientIdentifierType main, PatientIdentifierType toBeMerged) throws DAOException {
+		String nameOfMain = main.getName();
+		String nameOfToBeMerged = toBeMerged.getName();
+		
+		// Check if there are duplicate patient identifiers. Throw DAOException if there is
+		String checkDuplicates = "SELECT  COUNT(identifier) FROM patient_identifier WHERE identifier_type = :mainName or identifier_type = :mergedName GROUP BY identifier HAVING COUNT(identifier) > 1";
+		SQLQuery queryDuplicates = sessionFactory.getCurrentSession().createSQLQuery(checkDuplicates);
+		queryDuplicates.setParameter("mainName", main.getPatientIdentifierTypeId());
+		queryDuplicates.setParameter("mergedName", toBeMerged.getPatientIdentifierTypeId());
+		if (queryDuplicates.list().size() >= 1) {
+			throw new DAOException("Can't automatically merge. Duplicate patient identifiers");
+		}
+		
+		// Merge the patient identifiers of nameOfMain and nameOfToBeMerged
+		String queryString = "UPDATE patient_identifier SET identifier_type = (SELECT patient_identifier_type_id FROM patient_identifier_type WHERE NAME = :mainName) " + 
+				"WHERE identifier_type = (SELECT patient_identifier_type_id FROM patient_identifier_type WHERE NAME = :mergedName)";
+		SQLQuery query = sessionFactory.getCurrentSession().createSQLQuery(queryString);
+		query.setParameter("mainName", nameOfMain);
+		query.setParameter("mergedName", nameOfToBeMerged);
+		query.executeUpdate();
+
+		// Delete the patientIdentifierType that was merged
+		String deleteString = "DELETE FROM patient_identifier_type WHERE NAME = :mergedName";
+		SQLQuery deleteQuery = sessionFactory.getCurrentSession().createSQLQuery(deleteString);
+		deleteQuery.setParameter("mergedName", nameOfToBeMerged);
+		deleteQuery.executeUpdate();
+		}
 }
